Started Parent views for Main Parent Page and Update Child Page
Goal: Have it so students of logged in parent are listed with a hyperlink to their update student info page. The update student info page will allow changes to whether a student should be tracked by the system and the students picture image.

File upload button and a checkbox for tracking is implemented into the updateC.html, but has no current use for django.

Issues for logging in a parent, knowing who is logged in and displaying the proper list of students.