Pages for Parent and Update Child/Student Accessible but not functional
Need to be able to know which parent user is currently logged in, in order to properly query the Student model and display all students associated with that parent.

Need to be able to use the input fields in updateC.html to take input and affect the specific Student of the logged in parent that the parent has clicked in the main parent page (The specific student that they want to adjust information for).