Updated Card, Table and App
Card can now communicate with the backend. Definitely not final. Needs to display proper information, still using dummy values.

SimpleTable can display every event and relevant information that can be found in the database of the backend.

App.js updated to include paths to the parentview and events view created.