Updated Serializers.py
Update was made to include the id fields of database entries. This needed to be done because specific fields of the models are references to these id fields.

Makes it easier to find and search through the models on the front end if these fields are returned as a part of the json objects.