Updated Events, ParentView, and SimpleCard
Made it so SimpleCard can take passed in values and display them.

ParentView not finished, needs to be able to know what parent is signed in, their children/students, the most recent event for each child, the bus the event took place on, when the event happened, and who was driving the bus.

That information needs to be displayed on a card (pass that info into SimpleCard as done with dummy values in ParentView)

Events displays every event in the database. Defaults to Events for All Events. Needs to be able to fill selection box with logged in parents students and have it so text changes with what is selected in the selection box.